+++ x/simulation/simulate.go
@@ -105,6 +105,10 @@ func SimulateFromSeed(
 
 	accs = tmpAccs
 	nextValidators := validators
+	if len(nextValidators) == 0 {
+		tb.Skip("skipping: empty validator set in genesis")
+		return true, params, nil
+	}
 
 	var (
 		pastTimes          []time.Time
@@ -243,6 +247,10 @@ func SimulateFromSeed(
 		// on the next block
 		validators = nextValidators
 		nextValidators = updateValidators(tb, r, params, validators, res.ValidatorUpdates, eventStats.Tally)
+		if len(nextValidators) == 0 {
+			tb.Skip("skipping: empty validator set")
+			return true, params, nil
+		}
 
 		// update the exported params
 		if config.ExportParamsPath != "" && int64(config.ExportParamsHeight) == blockHeight {